move UnsignedByteArraySerializer and deserializer into model.jackson.modules and add to ObjectMapper.
