+++ xrpl4j-core/src/main/java/org/xrpl/xrpl4j/crypto/core/signing/Signature.java
@@ -52,8 +52,6 @@ public interface Signature {
    *
    * @return A {@link UnsignedByteArray}.
    */
-  @JsonSerialize(using = UnsignedByteArraySerializer.class)
-  @JsonDeserialize(using = UnsignedByteArrayDeserializer.class)
   UnsignedByteArray value();
 
   /**
